Move pollset to default multiplexer
